+++ k8s.tf
@@ -1,5 +1,3 @@
-
-
 module "eks" {
   source                                 = "git::https://github.com/aws-ia/terraform-aws-eks-blueprints.git?ref=v4.32.1"
   cluster_version                        = var.kubernetesVersion
@@ -13,52 +11,20 @@ module "eks" {
   tags                                   = var.tags
   cloudwatch_log_group_kms_key_id        = aws_kms_key.kms_key_cloudwatch_log_group.arn
   cloudwatch_log_group_retention_in_days = var.cloudwatch_retention
-  managed_node_groups                    = merge(local.default_node_pool, local.exec_node_pools, var.gpuNodePool ? local.gpu_node_pool : {}, var.ivsGpuNodePool ? local.ivsgpu_node_pool : {})
+  managed_node_groups                    = merge(local.default_managed_node_pools, local.exec_node_pools, var.gpuNodePool ? local.gpu_node_pool : {}, var.ivsGpuNodePool ? local.ivsgpu_node_pool : {})
 }
 
+data "aws_eks_node_group" "default" {
+  cluster_name    = local.infrastructurename
+  node_group_name = replace(module.eks.managed_node_groups[0]["default"]["managed_nodegroup_id"][0], "${local.infrastructurename}:", "")
 
-module "eks-addons" {
-  source                               = "git::https://github.com/aws-ia/terraform-aws-eks-blueprints.git//modules/kubernetes-addons?ref=v4.32.1"
-  eks_cluster_id                       = module.eks.eks_cluster_id
-  enable_amazon_eks_vpc_cni            = true
-  enable_amazon_eks_coredns            = true
-  enable_amazon_eks_kube_proxy         = true
-  enable_aws_efs_csi_driver            = true
-  enable_amazon_eks_aws_ebs_csi_driver = true
-  enable_aws_load_balancer_controller  = false
-  enable_cluster_autoscaler            = true
-  enable_aws_for_fluentbit             = var.enable_aws_for_fluentbit
-  enable_ingress_nginx                 = var.enable_ingress_nginx
-  tags                                 = var.tags
-  aws_for_fluentbit_helm_config = {
-    values = [templatefile("${path.module}/templates/fluentbit_values.yaml", {
-      aws_region           = data.aws_region.current.name,
-      log_group_name       = local.log_group_name,
-      service_account_name = "aws-for-fluent-bit-sa"
-    })]
-    dependency_update = true
-  }
-
-  ingress_nginx_helm_config = {
-    values = [templatefile("${path.module}/templates/nginx_values.yaml", {
-      internal       = "false",
-      scheme         = "internet-facing",
-      public_subnets = join(", ", local.public_subnets)
-    })]
-    namespace         = "nginx",
-    create_namespace  = true
-    dependency_update = true
-  }
-
-  cluster_autoscaler_helm_config = var.cluster_autoscaler_helm_config
-  #depends_on                     = [module.eks.managed_node_groups]
 }
-
 
 data "aws_eks_node_group" "execnodes" {
   count           = length(var.team_names)
   cluster_name    = local.infrastructurename
   node_group_name = replace(module.eks.managed_node_groups[0][var.team_names[count.index]]["managed_nodegroup_id"][0], "${local.infrastructurename}:", "")
+
 }
 
 data "aws_eks_node_group" "gpuexecnodes" {
@@ -73,6 +39,17 @@ data "aws_eks_node_group" "gpuivsnodes" {
   node_group_name = replace(module.eks.managed_node_groups[0]["gpuivsnodes"]["managed_nodegroup_id"][0], "${local.infrastructurename}:", "")
 }
 
+resource "aws_autoscaling_group_tag" "default_node-template_resources_ephemeral-storage" {
+  autoscaling_group_name = data.aws_eks_node_group.default.resources[0].autoscaling_groups[0].name
+
+  tag {
+    key   = "k8s.io/cluster-autoscaler/node-template/resources/ephemeral-storage"
+    value = "${var.linuxNodeDiskSize}G"
+
+    propagate_at_launch = true
+  }
+}
+
 resource "aws_autoscaling_group_tag" "execnodes" {
   count                  = length(var.team_names)
   autoscaling_group_name = data.aws_eks_node_group.execnodes[count.index].resources[0].autoscaling_groups[0].name
@@ -85,6 +62,19 @@ resource "aws_autoscaling_group_tag" "execnodes" {
   }
 }
 
+# see https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/cloudprovider/aws/README.md#auto-discovery-setup
+#     https://github.com/kubernetes/autoscaler/issues/1869#issuecomment-518530724
+resource "aws_autoscaling_group_tag" "execnodes_node-template_resources_ephemeral-storage" {
+  autoscaling_group_name = data.aws_eks_node_group.execnodes.resources[0].autoscaling_groups[0].name
+
+  tag {
+    key   = "k8s.io/cluster-autoscaler/node-template/resources/ephemeral-storage"
+    value = "${var.linuxExecutionNodeDiskSize}G"
+
+    propagate_at_launch = true
+  }
+}
+
 resource "aws_autoscaling_group_tag" "gpuexecnodes" {
   count                  = var.gpuNodePool ? 1 : 0
   autoscaling_group_name = data.aws_eks_node_group.gpuexecnodes[0].resources[0].autoscaling_groups[0].name
@@ -92,6 +82,18 @@ resource "aws_autoscaling_group_tag" "gpuexecnodes" {
   tag {
     key   = "k8s.io/cluster-autoscaler/node-template/label/purpose"
     value = "gpu"
+
+    propagate_at_launch = true
+  }
+}
+
+resource "aws_autoscaling_group_tag" "gpuexecnodes_node-template_resources_ephemeral-storage" {
+  count                  = var.gpuNodePool ? 1 : 0
+  autoscaling_group_name = data.aws_eks_node_group.gpuexecnodes[0].resources[0].autoscaling_groups[0].name
+
+  tag {
+    key   = "k8s.io/cluster-autoscaler/node-template/resources/ephemeral-storage"
+    value = "${var.gpuNodeDiskSize}G"
 
     propagate_at_launch = true
   }