+++ locals.tf
@@ -3,7 +3,7 @@ data "aws_ami" "al2gpu_ami" {
   most_recent = true
   filter {
     name   = "name"
-    values = ["amazon-eks-gpu-node-1.31-v20241109"]
+    values = ["*amazon-eks-gpu-node-${var.kubernetesVersion}*"]
   }
 }
 
@@ -13,7 +13,6 @@ locals {
   use_private_subnets_ids                   = length(var.private_subnet_ids) == 0 ? false : true
   use_public_subnet_ids                     = length(var.public_subnet_ids) == 0 ? false : true
   infrastructurename                        = var.infrastructurename
-  log_group_name                            = "/${module.eks.eks_cluster_id}/worker-fluentbit-logs"
   account_id                                = data.aws_caller_identity.current.account_id
   region                                    = data.aws_region.current.name
   license_server_role                       = "${local.infrastructurename}-license-server-role"
@@ -26,14 +25,14 @@ locals {
   patch_manager_cloudwatch_loggroup_install = "/aws/ssm/${module.eks.eks_cluster_id}/install"
   patchgroupid                              = "${var.infrastructurename}-patch-group"
   s3_instance_buckets                       = flatten([for name, instance in module.simphera_instance : instance.s3_buckets])
-  #license_server_bucket                     = var.licenseServer ? [aws_s3_bucket.license_server_bucket[0].bucket] : []
-  s3_buckets      = concat(local.s3_instance_buckets, [aws_s3_bucket.bucket_logs.bucket]) #, local.license_server_bucket)
-  private_subnets = var.private_subnet_ids                                                # local.create_vpc ? module.vpc[0].private_subnets : (local.use_private_subnets_ids ? var.private_subnet_ids : [for s in data.aws_subnet.private_subnet : s.id])
-  public_subnets  = var.public_subnet_ids                                                 #local.create_vpc ? module.vpc[0].public_subnets : (local.use_public_subnet_ids ? var.public_subnet_ids : [for s in data.aws_subnet.public_subnet : s.id])
+  license_server_bucket                     = var.licenseServer ? [aws_s3_bucket.license_server_bucket[0].bucket] : []
+  s3_buckets                                = concat(local.s3_instance_buckets, [aws_s3_bucket.bucket_logs.bucket]) #, local.license_server_bucket)
+  private_subnets                           = var.private_subnet_ids                                                # local.create_vpc ? module.vpc[0].private_subnets : (local.use_private_subnets_ids ? var.private_subnet_ids : [for s in data.aws_subnet.private_subnet : s.id])
+  public_subnets                            = var.public_subnet_ids                                                 # local.create_vpc ? module.vpc[0].public_subnets : (local.use_public_subnet_ids ? var.public_subnet_ids : [for s in data.aws_subnet.public_subnet : s.id])
   # Using a one-line command for gpuPostUserData to avoid issues due to different line endings between Windows and Linux.
   gpuPostUserData = "sudo yum -y erase nvidia-driver \nsudo yum -y install make gcc \nsudo yum -y update \nsudo yum -y install gcc kernel-devel-$(uname -r) \nsudo curl -fSsl -O https://us.download.nvidia.com/tesla/${var.gpuNvidiaDriverVersion}/NVIDIA-Linux-x86_64-${var.gpuNvidiaDriverVersion}.run \nsudo chmod +x NVIDIA-Linux-x86_64*.run \nsudo CC=/usr/bin/gcc10-cc ./NVIDIA-Linux-x86_64*.run -s --no-dkms --install-libglvnd \nsudo touch /etc/modprobe.d/nvidia.conf \necho \"options nvidia NVreg_EnableGpuFirmware=0\" | sudo tee --append /etc/modprobe.d/nvidia.conf \nsudo reboot"
 
-  default_node_pool = {
+  default_managed_node_pools = {
     "default" = {
       node_group_name = "default"
       instance_types  = var.linuxNodeSize
@@ -41,9 +40,9 @@ locals {
       desired_size    = var.linuxNodeCountMin
       max_size        = var.linuxNodeCountMax
       min_size        = var.linuxNodeCountMin
+      disk_size       = var.linuxNodeDiskSize
     }
   }
-
   exec_node_pools = {
     for team_name in var.team_names :
     "${team_name}" => {